You can upload files and download your invoices

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const Product = require('../models/product');
 const {validationResult}=require('express-validator');
 const { createSecretKey } = require('crypto');
+const fileHelper=require('../util/Dfile');
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/add-product', {
     pageTitle: 'Add Product',
@@ -18,11 +19,10 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const errors=validationResult(req);
-  console.log(errors);
   if(!errors.isEmpty()){
     return res.status(422).render('admin/add-product', {
       pageTitle: 'Add Product',
@@ -30,12 +30,12 @@ exports.postAddProduct = (req, res, next) => {
       errorMessage:errors.array()[0].msg,
       oldInput:{
         title:title,
-        imageUrl:imageUrl,
         price:price,
         description:description,
       }
     });
   }
+  const imageUrl=image.path;
   const product = new Product({
     title: title,
     price: price,
@@ -52,6 +52,7 @@ exports.postAddProduct = (req, res, next) => {
     })
     .catch(err => {
       const error= new Error(err);
+      console.log(error);
       error.httpStatusCode=500;
       return next(error);
     });
@@ -76,7 +77,6 @@ exports.getEditProduct = (req, res, next) => {
         errorMessage:'',
         oldInput:{
           title:'',
-          imageUrl:'',
           price:'',
           description:'',
         }
@@ -84,6 +84,7 @@ exports.getEditProduct = (req, res, next) => {
     })
     .catch(err => {
       const error= new Error(err);
+      console.log(error);
       error.httpStatusCode=500;
       return next(error);
     });
@@ -93,7 +94,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
   const errors=validationResult(req);
   
@@ -110,7 +111,6 @@ exports.postEditProduct = (req, res, next) => {
           product:{
             title:updatedTitle,
             price:updatedPrice,
-            imageUrl:updatedImageUrl,
             description:updatedDesc,
             _id:prodId
           }
@@ -119,7 +119,9 @@ exports.postEditProduct = (req, res, next) => {
         product.title=updatedTitle;
         product.price=updatedPrice;
         product.description=updatedDesc;
-        product.imageUrl=updatedImageUrl;
+        if(image){
+          product.imageUrl=image.path;
+        }
         product.userId=req.session.user._id;
         return product.save().then(result => {
           console.log('UPDATED PRODUCT!');
@@ -155,8 +157,13 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id:prodId,userId:req.user._id})
-    .then(() => {
+  Product.findById(prodId).then(product=>{
+    if(!product){
+      return next(new Error('Product not found'));
+    }
+    fileHelper.delecteFile(product.imageUrl);
+    return Product.deleteOne({_id:prodId,userId:req.user._id});
+  }).then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
     })
